refactor: moves context utility to factories module
Relocates the context creation helper function to improve code organization and module cohesion.

Simplifies the function signature by removing the optional parameter from the getter function while maintaining the same core functionality.
