+++ src/lib/util.ts
@@ -3,7 +3,7 @@ import { wait } from "@sillvva/utils";
 import type { StandardSchemaV1 } from "@standard-schema/spec";
 import { hotkey as hk, type HotkeyItem } from "@svelteuidev/composables";
 import { isObject } from "effect/Predicate";
-import { getContext, hasContext, setContext, tick } from "svelte";
+import { tick } from "svelte";
 import type { Attachment } from "svelte/attachments";
 import type { FullPathname } from "./constants";
 import { errorToast } from "./factories.svelte";
@@ -82,19 +82,6 @@ export function getRelativeTime(date: Date | number, lang = navigator.language):
 
 export function isStandardSchema(schema: unknown): schema is StandardSchemaV1 {
 	return isObject(schema) && "~standard" in schema;
-}
-
-export function createContext<T>(createDefault?: () => T): [(getDefault?: () => T) => T, (context: T) => T] {
-	const key = Symbol("context");
-	return [
-		(getDefault) => {
-			if (hasContext(key)) return getContext(key);
-			if (getDefault) return setContext(key, getDefault());
-			if (createDefault) return setContext(key, createDefault());
-			throw new Error("Context not found");
-		},
-		(context) => setContext(key, context)
-	];
 }
 
 export function isRedirectFailure(
